pyproject.toml
[tool.poetry]
name = "mangals-backend"
version = "0.1.0"
description = ""
authors = ["Vadim <skolkovoinovations@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
